postgribbit/requirements.txt
Flask==2.1.1
requests==2.26.0
psycopg2-binary
werkzeug==2.3.7